+++ http/jakarta-json/pom.xml
@@ -5,11 +5,11 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>org.extism.sdk</groupId>
-        <artifactId>chicory-sdk</artifactId>
+        <artifactId>http</artifactId>
         <version>999-SNAPSHOT</version>
     </parent>
 
-    <artifactId>chicory-sdk-core</artifactId>
+    <artifactId>http-jakarta-json</artifactId>
 
     <properties>
         <maven.compiler.source>22</maven.compiler.source>
@@ -17,37 +17,23 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
 
+
     <dependencies>
-        <dependency>
-            <groupId>com.dylibso.chicory</groupId>
-            <artifactId>runtime</artifactId>
-            <version>${chicory.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.dylibso.chicory</groupId>
-            <artifactId>wasi</artifactId>
-            <version>${chicory.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.dylibso.chicory</groupId>
-            <artifactId>aot-experimental</artifactId>
-            <version>${chicory.version}</version>
-        </dependency>
         <dependency>
             <groupId>org.extism.sdk</groupId>
             <artifactId>http-api</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <version>${junit.version}</version>
-            <scope>test</scope>
+            <groupId>jakarta.json</groupId>
+            <artifactId>jakarta.json-api</artifactId>
+            <version>${jakarta.json-api.version}</version>
+            <optional>true</optional>
         </dependency>
         <dependency>
-            <groupId>com.google.jimfs</groupId>
-            <artifactId>jimfs</artifactId>
-            <version>${jimfs.version}</version>
+            <groupId>org.eclipse.parsson</groupId>
+            <artifactId>jakarta.json</artifactId>
+            <version>${jakarta.json.version}</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
